Fix invalid-symbol crash in LivenessToSafetyTranslator for quoted names
LivenessToSafetyTranslator built each "loop" state's name via
statevar->to_string() + suffix. to_string() already returns the
SMT-LIB-quoted form for names that aren't simple symbols (e.g.
hierarchical names containing '.' or '[]'), so appending more text
after it produced a malformed name like
"|mod.sig[3]|__pono_generated___loop" -- a quoted symbol with
trailing characters after its closing '|'. Bitwuzla rejects this with
"invalid symbol ..., symbol is not SMT-LIB compliant"; other backends
silently mangle it instead.

Strip a matching pipe-quote pair before appending, using
name_desanitize -- moved from utils/syntax_analysis_common.{h,cpp}
(IC3SA/SyGuS-specific, and coupled to sygus_ic3formula_helper.h) into
utils/str_util.{h,cpp}, which already holds general-purpose string
helpers (e.g. lookup_or_key) used outside that subsystem.

Adds a regression test using a state variable name shaped like a
SystemVerilog hierarchical path, which reproduces the exact reported
error against bzla before the fix.
